refactor: add exception for entity not found on update and delete endpoint

+++ src/main/java/dev/reso/practicing/service/MissionService.java
@@ -53,11 +53,15 @@ public class MissionService {
             return missionMapper.map(mission);
         }
 
-        return null;
+        throw new ResourceNotFoundException("Not found id -> " + id);
     }
 
     @Transactional
     public void delete(Long id){
-        missionRepository.deleteById(id);
+        if(missionRepository.findById(id).isPresent()){
+            missionRepository.deleteById(id);
+        }else{
+            throw new ResourceNotFoundException("Not Found id -> " + id);
+        }
     }
 }
